Update test.ps1 script to break fast on test failures

+++ script/test.ps1
@@ -8,6 +8,14 @@ Start-AzureFunction 7074 -workingDir "test\Indigo.Functions.Injection.Integratio
 
 dotnet test test\Indigo.Functions.Configuration.IntegrationTests\Indigo.Functions.Configuration.IntegrationTests.csproj
 
+if ($LastExitCode -ne 0) {
+    return $LastExitCode
+}
+
 dotnet test test\Indigo.Functions.Injection.IntegrationTests.InjectionTests\Indigo.Functions.Injection.IntegrationTests.InjectionTests.csproj
+
+if ($LastExitCode -ne 0) {
+    return $LastExitCode
+}
 
 Stop-Process (Get-Process func).Id
